Delete original copy of at_cmd_parser.
This module has been moved to fw-nrfconnect-nrf, and has
been taken in use there. This copy complicates use of the one
in that repo as headers are included in two paths, and not
matching anymore. This patch deletes the original copy
and mandates users to move to the new module.

Currently in this module is unused in this repository.