NCSDK-2286: Update Manifest.
Get a more recent version of fw-nrfconnect-nrf to use new modules. It
also require to update fw-nrfconnect-zephyr.
CoAP fixes are now part of the LwM2M master branch.
Get bsdlib 0.3.1 with latest fixes and APN support.
MbedTLS is now a required dependency in order to build the nrf_security
module available in nrfxlib.

Signed-off-by: Christopher Métrailler <christopher.metrailler@nordicsemi.no>