lwm2m_firmware_download: handle case where scratch area is not erased
It could happen, after a manual or specific firmware update,
that the scratch area is not erased even though the offset
reported by the modem is zero.
In that case, LwM2M stack would start the download but fail.
Handle that case by re-starting the download task, which
will this time erase the scratch area, since it would then
be properly marked as "dirty".

Print RPC error when modem has rejected a firmware fragment.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>