lwm2m_firmware_download: reinit DFU socket when download is completed
Re-initialize the DFU socket when the download is completed to
guard against incorrect server behavior, where a new download
is started without the previous one being applied.

In that situation the DFU socket would be closed, and the
download_task would be stuck in loop waiting for the firmware
image in flash to be deleted.

Instead, re-initialize the DFU socket so that if the server
prompt us to download a new firmware image, the download_task
can fetch the offset and correctly determine that a firmware
image is already present, and just skip the download.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>