lwm2m_firmware_download: remember erase state when downloading
It could happen that power is lost while the modem is erasing
its scratch area to make space for the new firmware.

This patch extends the state information saved by the stack
to detect this condition and resumes the update process
on boot if necessary.

The signature of lwm2m_firmware_image_ready_set/get() has been
changed to lwm2m_firmware_image_state_set()/get() and the logic
in lwm2m_firmware_download_init() has been simplified as
a consequence.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>