NCSDK-NONE: Do not assert if MSISDN is not available.
This handle a specific case in the client application.
The debug print shell command can be used to print the MSISDN value.
MSISDN is read from the SIM only when connected to the network. If we
call this shell command before being connected, we will now print an
empty value instead of asserting.

Signed-off-by: Christopher Métrailler <christopher.metrailler@nordicsemi.no>