lwm2m_firmware_download: handle invalid content ranges in response
When we have a non-zero, non-dirty offset but the server won't
send any more bytes, so we can't continue the download.
We likely lost power after downloading the whole image but
before we could save that information in flash.
Manually override the image state to attempt to apply
the existing patch.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>