Initiate bootstrap after last retry delay
The latest VzW requirements states that if the fourth retry attempt
is unsuccessful, the device shall fail the session and repeat the
bootstrap flow after 24 hours or after a reboot of the device.

In case registration with the LwM2M management server fails because
it receives a 4.00 error, the device shall fail the session and not
attempt additional retry for 24 hours.