Remove APP_ERROR_CHECK and APP_ERROR_CHECK_BOOL
For lwm2m_register() failure close the connection and reconnect in
a hope that the issue will be fixed.

For lwm2m_deregister() failure log the error and continue.
This should be handled in a separate "deregister reconnect" fix.

For coap_init() failure return error from lwm2m_carrier_init().