LwM2M Client refactoring: observer process
Modified lwm2m_firmware_observer_process so that it behaves in a similar
manner as lwm2m_conn_mon_observer_process, so that observer notifications
can be sent to a particular remote upon reconnection if it failed to send.
Got rid of code duplication, as lwm2m_firmware_observer_process was a copy
of lwm2m_firmware_notify_resource applied to a single resource.

Modified app_lwm2m_observer_process to enable sending notifications after
reconnection to all the observers (the ones specified in the observer_process),
and not only the ones corresponding to the Connectivity Monitoring object.