lwm2m_vzw: turn off the link if erase times out
Network behavior affects whether or not the modem can erase
the delta image in its flash. Turn off the link if the modem
takes too long to erase the delta image in flash.
Once the erase has happened, turn on the link again.

Added two events:
- LWM2M_CARRIER_EVENT_CONNECTING
- LWM2M_CARRIER_EVENT_DISCONNECTING

that are sent from the lib to the application when the library
is about to turn on and off the link respectively, to inform
of the imminent network state change.

Added a print statement in the weak library event handler
implementation to better see when we are sending events
and which events we are sending.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>