NCSDK-3752: Do not build the nrf client app by default
The nrf client application is not compiled by default anymore. When updating the
OS abstraction layer of the LWM2M library, the client application compilation
may fail until the abstraction layer has been fixed upstream in -nrf. This
process may take some time and a red build does not help us adding new features.

The nrf client application is not compiled by default anymore. This parameter
can be changed for each run of the pipeline. The CI in -nrf is now used to build
the nrf client application.

Signed-off-by: Christopher Métrailler <christopher.metrailler@nordicsemi.no>