LWM2M Notification Attribute and Write-Attribute Support
NCSDK-2942: Notification Attribute support

Implemented notification attributes (p_min, p_max, gt, lt and st) that
provide parameters to the "Notify" operation (and hence determine whether
a notification should be sent or not).

The notification attributes are stored in an array of size
LWM2M_MAX_NOTIF_ATTRIBUTE_TYPE within the lwm2m_observable_metadata_t
structure, which describes an individual observable item/resource to enable
easier management of observer notifications. The lwm2m_notif_attribute_t
structure stores information regarding the value of the corresponding
attribute, as well as its assignment level, since it is necessary to keep
track of it given the precedence rules that need to be applied during the
WRITE-ATTRIBUTE operation.

Modified the observer process, as it now iterates the list of the
initialized lwm2m_observable_metadata_t structures (which are always
initialized upon observer registration).

The notification attributes are stored in flash memory alongside their
corresponding observer, hence they are restored after reboot. The entry is
updated every time the attributes change their value or their assignation
level.

NCSDK-1267: LWM2M Write-Attribute support

Implemented a generic handler function for the incoming write-attribute
requests. Added callback functions to handle requests on objects (for now
only the WRITE operation on the objects with observable resources).

The handler identifies the specified notification attributes to be set or
unset (if no value given) and updates the corresponding observable items of
the given observer, if the precedence rules are respected and the coherency
checks are successful.

Signed-off-by: Kacper Radoszewski <kacper.radoszewski@nordicsemi.no>