LWM2M Client: Observer handling refactoring
NCSDK-4716: Support observers on object and object instance level
The current implementation of the observers in the library does not allow
registering, notifying nor storing observers that are not set at resource
level; however, one of the tests in AT&T requires that we support an
observation of an entire object. Modified the assignment of the
`resource_of_interest`, which is now a pointer to the observable item
itself, and not a resource ID, as it was previously done.

NCSDK-4717: Observer storage in flash
Changed the representation of the observers in NVS, as they are now
identified with the URI path of the observable item.

Split the creation of CoAP messages into several functions, so that they
can be constructed progressively and can be reused for the different types
of messages sent to the server.

Signed-off-by: Kacper Radoszewski <kacper.radoszewski@nordicsemi.no>