LWM2M Client: optimize observer notification type handling
Removed the object specific tables that keep track of the time elapsed from
the last time an observer notification has been sent as a CoAP CON message,
as it is now being tracked individually for each observable via a new field
added to the lwm2m_observable_metadata_t structure. This solution will save
some RAM as the timestamps are only stored for the initialized observables,
and it will also make it possible to use the same logic for all the
items, not just resources.

Signed-off-by: Kacper Radoszewski <kacper.radoszewski@nordicsemi.no>