NCSDK-4745: Generic mechanism to send observer notifications
Implemented a generic function to send the observer notifications given the
URI path of the item of interest of the observer, and removed the object
specific "lwm2m_xxx_notify_resource()" functions, as the only fundamental
difference between these was the object-specific TLV encoding, hence also
implemented an API to perform the TLV encoding on any observable, given its
URI path. This will also make it possible to send notifications to
observers at object or object instance level, and not solely resource.

Signed-off-by: Kacper Radoszewski <kacper.radoszewski@nordicsemi.no>