NCSDK-4545: enable setting a custom APN by the application
The application can now set a custom APN through the lwm2m_carrier_config_t
struct passed to lwm2m_carrier_init(), which will initialize an additional
instance of the APN Connection Profile object, if operating in the AT&T
network. Added a shell command to perform the same operation for the ease
of testing.

Signed-off-by: Kacper Radoszewski <kacper.radoszewski@nordicsemi.no>