lwm2m: improve obfuscation by obfuscating whole lib
Instead of obfuscating only nrf_lwm2m_carrier,
now obfuscate the whole library, that is merged
from nrf_coap, nrf_lwm2m and nrf_lwm2m_carrier.

This means that we won't have to maintain a list
of functions that are in nrf_coap and nrf_lwm2m
that we don't want to obfuscate.

Additionally, use a regexp to avoid aving to update
the filter list every time we add a lwm2m_carrier
or lwm2m_os API.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>