lwm2m_firmware_download: do to attempt to re-download on errors
When an image is rejected by the modem, we should not download it
again on boot; a failing image will still be a failing image.

When we fail the download, we should not attempt it again on boot.
We have plenty of logic to retry on failures, but when we give up
we must not try again on boot.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>