NCSDK-6210: Simplify West manifest and use imports
Simplifying the LwM2M West manifest. Modules dependencies and versions are now
dictated by the West manifest in -nrf. Modules versions are now updated
automatically when the -nrf revision changes.

West version 0.7 is required to use the 'import' feature.
Modules can be imported or not using a whitelist or blacklist.

Specific module versions can be overridden to test specific versions or new
features for instance. The override file could be created automatically in CI.

Signed-off-by: Christopher Métrailler <christopher.metrailler@nordicsemi.no>