lwm2m: in-band FOTA and other updates
Added support for in-band FOTA.

With this patch we move to a lazy erasing of the modem firmware
(only when a new one is being download) to a no-so-lazy policy where
we erase it if the modem rejects it, or during boot if we have just
updated. This is necessary because the device can now be pushed a
new firmware at any time when connected to the server, so the scratch
area must be ready to receive it at any time. Since it takes a long
time to do the erase, and the device may need to go offline to erase
in a timely manner, we can't start erasing if the firmware is incoming.

Signed-off-by: Emanuele Di Santo <emdi@nordicsemi.no>