NCSDK-6476: Change event READY behaviour
Change event READY to be given when LTE link is up and no bootstrap
sequence will be done. This will give the application an indication
when to start doing their logic a bit earlier.

Add a new event REGISTERED which will be given when all servers are
registered (where READY used to be given).

Moved event BOOTSTRAPPED to before writing keys, because the
boostrap sequence is done from the server side.