NCSDK-5000: Improve avoiding AT&T specific objects
Don't remove objects when they are not initialized, because then
the object_id is 0 and this will delete the security object.

Setup operator-specific objects when carrier change is detected,
because this may happen after a SIM swap (or when debugging).

Check if object already exists before adding.

Improve loop for deleting object instances.

Increase maximum number of instances because this may be needed
while switching operator.

Signed-off-by: Stig Bjørlykke <stig.bjorlykke@nordicsemi.no>