NCSDK-6615: Postpone storing is-bootstrapped for VzW
When doing bootstrap in VzW in certification mode (using Motive)
the "is bootstrapped" setting is not stored before successfully
registered with all servers.

This will avoid the need to remove all user available non-volatile
memory when the "Create Device Bootstrap" task is not executed or
fails, because the bootstrap sequence will be started at next restart.

Signed-off-by: Stig Bjørlykke <stig.bjorlykke@nordicsemi.no>